added the --pyplot option:
It will produce a plot with all cotranscriptional folding trajectories,
kind of an alternative to DrForna... In order to map the *ID* to a
secondary structure, use the --verbose option.

The plotting parameters may not be ideal, but they very much depend
on the problem and what you want to visualize.